Rename peeler/pyproejct/utils.py to parser.py
